editor/editor_path.cpp

+++ editor/editor_path.cpp
@@ -123,7 +123,7 @@ void EditorPath::update_path() {
 
 void EditorPath::_id_pressed(int p_idx) {
 
-	ERR_FAIL_INDEX(p_idx, objects.size());
+	ERR_FAIL_INDEX(p_idx, static_cast<int>(objects.size()));
 
 	Object *obj = ObjectDB::get_instance(objects[p_idx]);
 	if (!obj)
